added role checks to AccountController

+++ API/Controllers/AccountController.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using boxinator.Models;
 using boxinator.Models.Domain;
 using boxinator.Models.DTO.User;
 using boxinator.Services.Interfaces;
@@ -29,13 +30,22 @@ namespace boxinator.Controllers
         /// Get user information by id
         /// </summary>
         /// <param name="accountId"></param>
-        /// <returns>Retrieved user</returns>
+        /// <returns>Retrieved user or 403</returns>
         // GET: /account/:account_id
         [HttpGet("/account/{accountId}")]
         public async Task<ActionResult<UserReadDTO>> Get(int accountId)
         {
-            var resultUser = await _service.Get(accountId);
-            return _mapper.Map<UserReadDTO>(resultUser);
+            // current user for role check
+            var userEmail = Request.ExtractEmailFromToken();
+            User currentUser = await _service.GetUser(userEmail);
+
+            if (currentUser.IsAdmin() || currentUser.Id == accountId)
+            {
+                var resultUser = await _service.Get(accountId);
+                return _mapper.Map<UserReadDTO>(resultUser);
+            }
+
+            return StatusCode(403);
         }
 
         /// <summary>
@@ -43,54 +53,94 @@ namespace boxinator.Controllers
         /// </summary>
         /// <param name="accountId"></param>
         /// <param name="userDTO"></param>
-        /// <returns>Updated user</returns>
+        /// <returns>Updated user or 403</returns>
         // PUT: /account/:account_id
         [HttpPut("/account/{accountId}")]
         public async Task<ActionResult<UserReadDTO>> Update(int accountId, UserEditDTO userDTO)
         {
-            User updatedUser = _mapper.Map<User>(userDTO);
-            User resultUser = await _service.Update(accountId, updatedUser);
-            return _mapper.Map<UserReadDTO>(resultUser);
+            // current user for role check
+            var userEmail = Request.ExtractEmailFromToken();
+            User currentUser = await _service.GetUser(userEmail);
+
+            if(currentUser.IsAdmin() || currentUser.Id == accountId)
+            {
+                User updatedUser = _mapper.Map<User>(userDTO);
+                User resultUser = await _service.Update(accountId, updatedUser);
+                return _mapper.Map<UserReadDTO>(resultUser);
+            }
+
+            return StatusCode(403);
+
         }
 
         /// <summary>
         /// Add new user
         /// </summary>
         /// <param name="userDTO"></param>
-        /// <returns></returns>
+        /// <returns>Added user or 403</returns>
         // POST: /account
         [HttpPost]
         public async Task<ActionResult<UserReadDTO>> Add(UserCreateDTO userDTO)
         {
-            User newUser = _mapper.Map<User>(userDTO);
-            User resultUser = await _service.Add(newUser);
-            return _mapper.Map<UserReadDTO>(resultUser);
+            // current user for role check
+            var userEmail = Request.ExtractEmailFromToken();
+            User currentUser = await _service.GetUser(userEmail);
+
+            if (currentUser.IsAdmin())
+            {
+                User newUser = _mapper.Map<User>(userDTO);
+                User resultUser = await _service.Add(newUser);
+                return _mapper.Map<UserReadDTO>(resultUser);
+            }
+
+            return StatusCode(403);
+
         }
 
         /// <summary>
         /// Delete user
         /// </summary>
         /// <param name="accountId"></param>
-        /// <returns></returns>
+        /// <returns>200, 404 or 403</returns>
         // DELETE: /account/:account_id
         [HttpDelete("/account/{accountId}")]
-        public async Task<ActionResult<bool>> Delete(int accountId)
+        public async Task<ActionResult> Delete(int accountId)
         {
-            // RESTRICT TO ADMIN ONLY!
-            return await _service.Delete(accountId);
+            // current user for role check
+            var userEmail = Request.ExtractEmailFromToken();
+            User currentUser = await _service.GetUser(userEmail);
+
+            if (currentUser.IsAdmin())
+            {
+                bool success = await _service.Delete(accountId);
+                return success == true ? Ok() : StatusCode(404);
+            }
+
+            return StatusCode(403);
+
         }
 
         /// <summary>
         /// Search user
         /// </summary>
-        /// <returns>Found user</returns>
+        /// <returns>Found user or 403</returns>
         [HttpGet]
         public async Task<ActionResult<List<UserReadDTO>>> Search()
         {
-            string searchTerm = HttpContext.Request.Query["searchTerm"].ToString();
+            // current user for role check
+            var userEmail = Request.ExtractEmailFromToken();
+            User currentUser = await _service.GetUser(userEmail);
 
-            var resultUsers = await _service.Search(searchTerm);
-            return _mapper.Map<List<UserReadDTO>>(resultUsers);
+            if (currentUser.IsAdmin())
+            {
+                string searchTerm = HttpContext.Request.Query["searchTerm"].ToString();
+
+                var resultUsers = await _service.Search(searchTerm);
+                return _mapper.Map<List<UserReadDTO>>(resultUsers);
+            }
+
+            return StatusCode(403);
+
         }
     }
 }
